A non-phony make target to avoid needless rebuilds
Make the source files a proper dependency for the make target.
Includes a tweak to the way that GO_FILES was set so sees all the
go files in the subdirectories.

With this change you'll get `make: Nothing to be done for 'build'`
when the existing binary is still fresh. (Of course you can still
force a rebuild if you want with `make clean build`.)